Rename openMalaria.cpp to main.cpp and simplify workfow (#327)
* move free checkpoint functions in checkpoint.h and .cpp

* change static to inline for create*model functions

* rename model/openMalaria.cpp to model/main.cpp
